Changes to Company

+++ src/flyace/Company.java
@@ -7,11 +7,14 @@ import java.util.ArrayList;
 //	Adds the Company to the ticket
 //	Sends the ticket to the Plane
 //		plane.putCustomer(Ticket);
+/**
+ * 
+ * @author Stefan Elmgren
+ */
 public class Company {
-
-    private String name;
-    private int money;
-    private double profit = 0.3;
+    private String name; // Company name
+    private double money;
+    private double profit = 0.3; 
     private ArrayList<PassengerPlane> planes = new ArrayList<PassengerPlane>();
 
     //******************* Constructors *****************************************
@@ -23,7 +26,7 @@ public class Company {
 
     }
 
-    public void addMoney(int mon) {
+    public void addMoney(double mon) {
         this.money += mon;
     }
 
@@ -55,7 +58,7 @@ public class Company {
         return name;
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 
@@ -67,7 +70,7 @@ public class Company {
         return planes;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         this.money = money;
     }
 
@@ -79,10 +82,4 @@ public class Company {
         pPlane.putCustomer(ticket);
         
     }
-
-    //Company
-    //	Selects the plane which is presently boarding
-    //	Adds the Company to the ticket
-    //	Sends the ticket to the Plane
-    //		plane.putCustomer(Ticket);
 }
